Se ha arreglado el error de la edicion en la parte de los medicamentos y se ha comenzado a trabajar en la parte de busqueda de ventas NOTA: Ejecutar en postgresql la funcion de editarinventario ya que ahi estaba el error y se reestructura una buena parte de ella

+++ src/modelo/Cliente_modelo.java
@@ -71,7 +71,7 @@ public class Cliente_modelo {
         try {
             if (conn.Conectar()) {
                 CallableStatement cmd = conn.getConnection().prepareCall("{ call obtenercliente(?) }");
-                cmd.setInt(1, pCliente.getId());
+                cmd.setLong(1, pCliente.getId());
                 if (cmd.execute()) {
                     ResultSet resultado = cmd.getResultSet();
                     while (resultado.next()) {
